scripts/requirements-news.txt
feedparser==6.0.11
python-dotenv==1.0.1
supabase==2.4.5
